Awscli handler (#61)
* adding object to handle awscli response,first phase. adding warning flags;


* fixing text to float conversion (ronen-fr #42) (#64)


* fix integer conversion; add compile-warning flag


* fix stdin mode; fix compile warnings

+++ include/s3select.h
@@ -1226,7 +1226,7 @@ void push_is_null_predicate::builder(s3select* self, const char* a, const char* 
   std::string token(a, b);
   bool is_null = true;
 
-  for(int i=0;i<token.size();i++)
+  for(size_t i=0;i<token.size();i++)
   {//TODO use other scan rules
     bsc::parse_info<> info = bsc::parse(token.c_str()+i, (bsc::str_p("is") >> bsc::str_p("not") >> bsc::str_p("null")) , bsc::space_p);
     if (info.full)
